refactor: Rename files tftp_session.* to tftpSession.* (u-t OK!)
Github issue #4
